# docs/swagger.yaml
basePath: /
definitions:
  handlers.ExecuteToolRequest:
    properties:
      parameters:
        additionalProperties: true
        type: object
    type: object
  handlers.ExecuteToolResponse:
    properties:
      error:
        example: ""
        type: string
      result: {}
      success:
        example: true
        type: boolean
      tool_name:
        example: search_contacts
        type: string
    type: object
  handlers.HealthResponse:
    properties:
      status:
        example: healthy
        type: string
      time:
        example: "2025-09-24T23:56:42+02:00"
        type: string
    type: object
  handlers.SendVoiceNoteResponse:
    properties:
      error:
        example: ""
        type: string
      filename:
        example: voice_note.ogg
        type: string
      recipient:
        example: 1234567890@s.whatsapp.net
        type: string
      success:
        example: true
        type: boolean
      timestamp:
        example: "2025-09-25T00:50:40+02:00"
        type: string
    type: object
  main.ListMessagesRequest:
    properties:
      chat_jid:
        example: 1234567890@s.whatsapp.net
        type: string
      limit:
        example: 50
        type: integer
      offset:
        example: 0
        type: integer
    type: object
  main.SearchContactsRequest:
    properties:
      query:
        example: John
        type: string
    type: object
  main.SendMessageRequest:
    properties:
      message:
        example: Hello World
        type: string
      recipient:
        example: 1234567890@s.whatsapp.net
        type: string
    type: object
  tools.Tool:
    properties:
      description:
        example: Search for contacts by name or phone number
        type: string
      inputSchema:
        additionalProperties: true
        type: object
      name:
        example: search_contacts
        type: string
    type: object
  tools.ToolsResponse:
    properties:
      tools:
        items:
          $ref: '#/definitions/tools.Tool'
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A WhatsApp MCP (Model Context Protocol) server providing WhatsApp functionality
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: WhatsApp MCP Server API
  version: "1.0"
paths:
  /api/list-messages:
    post:
      consumes:
      - application/json
      description: Retrieve messages from a specific WhatsApp chat
      parameters:
      - description: List messages request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.ListMessagesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: List of messages
          schema:
            additionalProperties: true
            type: object
      summary: List messages from a chat
      tags:
      - API
  /api/search-contacts:
    post:
      consumes:
      - application/json
      description: Search for WhatsApp contacts by name or phone number
      parameters:
      - description: Search contacts request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.SearchContactsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: List of contacts
          schema:
            additionalProperties: true
            type: object
      summary: Search for contacts
      tags:
      - API
  /api/send-message:
    post:
      consumes:
      - application/json
      description: Send a message to a WhatsApp contact or group
      parameters:
      - description: Send message request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/main.SendMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Message sent confirmation
          schema:
            additionalProperties: true
            type: object
      summary: Send a WhatsApp message
      tags:
      - API
  /api/send-voice-note:
    post:
      consumes:
      - multipart/form-data
      description: Upload and send an audio file as a WhatsApp voice message
      parameters:
      - description: Recipient JID
        in: formData
        name: recipient
        required: true
        type: string
      - description: Audio file (.ogg opus format)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Voice note sent successfully
          schema:
            $ref: '#/definitions/handlers.SendVoiceNoteResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Send a voice note via WhatsApp
      tags:
      - API
  /events:
    get:
      consumes:
      - text/event-stream
      description: Establishes a Server-Sent Events connection for real-time MCP communication
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE connection established
          schema:
            type: string
      summary: Server-Sent Events endpoint for MCP communication
      tags:
      - MCP
  /health:
    get:
      consumes:
      - application/json
      description: Returns the current health status of the server
      produces:
      - application/json
      responses:
        "200":
          description: Server is healthy
          schema:
            $ref: '#/definitions/handlers.HealthResponse'
      summary: Health check endpoint
      tags:
      - System
  /sse:
    get:
      consumes:
      - text/event-stream
      description: Establishes a Server-Sent Events connection for real-time MCP communication
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE connection established
          schema:
            type: string
      summary: Server-Sent Events endpoint for MCP communication
      tags:
      - MCP
  /tools:
    get:
      consumes:
      - application/json
      description: Returns a list of all available MCP tools with their schemas
      produces:
      - application/json
      responses:
        "200":
          description: List of available tools
          schema:
            $ref: '#/definitions/tools.ToolsResponse'
      summary: List available MCP tools
      tags:
      - MCP
  /tools/{tool_name}/execute:
    post:
      consumes:
      - application/json
      description: Execute any available MCP tool with the provided parameters
      parameters:
      - description: Name of the tool to execute
        in: path
        name: tool_name
        required: true
        type: string
      - description: Tool execution request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ExecuteToolRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Tool execution result
          schema:
            $ref: '#/definitions/handlers.ExecuteToolResponse'
        "404":
          description: Tool not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Tool execution error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Execute a specific MCP tool
      tags:
      - API
schemes:
- http
- https
swagger: "2.0"